matrix works

+++ Praktikum3/Praktikum3/Matrix.h
@@ -6,11 +6,11 @@
 class Matrix
 {
 private:
-	std::vector<std::vector<double>> m_Values;
+	std::vector<std::vector<int>> m_Values;
 
 public:
 	Matrix();
-	void CreateGeneratorMatrix(std::vector < std::vector<double>> values); // definition
+	void CreateGeneratorMatrix(std::vector < std::vector<int>> values); // definition
 	void CreateCanonicalMatrix(); // 1
